Fix cell voltage not updating in race and display mode
The signal was using int instead of float

+++ src/ViewLayer/RaceModeDisplay/RaceModeDashboardView.h
@@ -82,8 +82,8 @@ private slots:
     void packNetPowerReceived(double);
     void auxVoltageReceived(int);
     void packStateOfChargeReceived(double);
-    void lowCellVoltageReceived(int);
-    void averageCellVoltageReceived(int);
+    void lowCellVoltageReceived(float);
+    void averageCellVoltageReceived(float);
     void highTemperatureReceived(int);
     void averageTemperatureReceived(int);
